+++ packages/client/src/contexts/MUDContext.tsx
@@ -1,5 +1,5 @@
 import { createContext, ReactNode, useContext } from "react";
-import { SetupResult } from "./mud/setup";
+import { SetupResult } from "../mud/setup";
 
 const MUDContext = createContext<SetupResult | null>(null);
