+++ packages/client/src/utils/terrainTypes.ts
