+++ packages/client/src/utils/useKeyboardMovement.ts
@@ -1,5 +1,5 @@
 import { useEffect, useState } from "react";
-import { useMUD } from "../MUDContext";
+import { useMUD } from "../contexts/MUDContext";
 
 export const useKeyboardMovement = () => {
   const {
